Split Cypress helper file into multiple helpers for readability
