+++ cypress/helpers/helper-page.ts
@@ -1,0 +1,5 @@
+export const createTestPageAndOpenParagraphModal = (pageName: string) => {
+  cy.drupalLogin('/node/add/page');
+  cy.findByLabelText('Title').type(pageName);
+  cy.openParagraphsModal();
+};
